@@ -24,17 +24,17 @@
     "fix-vulnerabilities": "npx yarn-audit-fix && npx browserslist --update-db"
   },
   "dependencies": {
-    "@angular/animations": "^17.2.3",
-    "@angular/cdk": "^17.2.1",
-    "@angular/common": "^17.2.3",
-    "@angular/compiler": "^17.2.3",
-    "@angular/core": "^17.2.3",
-    "@angular/forms": "^17.2.3",
-    "@angular/material": "^17.2.1",
-    "@angular/platform-browser": "^17.2.3",
-    "@angular/platform-browser-dynamic": "^17.2.3",
-    "@angular/router": "^17.2.3",
-    "@gilsdav/ngx-translate-router": "^7.0.0",
+    "@angular/animations": "^17.3.2",
+    "@angular/cdk": "^17.3.2",
+    "@angular/common": "^17.3.2",
+    "@angular/compiler": "^17.3.2",
+    "@angular/core": "^17.3.2",
+    "@angular/forms": "^17.3.2",
+    "@angular/material": "^17.3.2",
+    "@angular/platform-browser": "^17.3.2",
+    "@angular/platform-browser-dynamic": "^17.3.2",
+    "@angular/router": "^17.3.2",
+    "@gilsdav/ngx-translate-router": "^7.1.0",
     "@gilsdav/ngx-translate-router-http-loader": "^2.0.0",
     "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "^8.0.0",
@@ -47,24 +47,24 @@
     "zone.js": "^0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.2.2",
-    "@angular-eslint/builder": "^17.2.1",
-    "@angular-eslint/eslint-plugin": "^17.2.1",
-    "@angular-eslint/eslint-plugin-template": "^17.2.1",
-    "@angular-eslint/schematics": "^17.2.1",
-    "@angular-eslint/template-parser": "^17.2.1",
-    "@angular/cli": "^17.2.2",
-    "@angular/compiler-cli": "^17.2.3",
+    "@angular-devkit/build-angular": "^17.3.2",
+    "@angular-eslint/builder": "^17.3.0",
+    "@angular-eslint/eslint-plugin": "^17.3.0",
+    "@angular-eslint/eslint-plugin-template": "^17.3.0",
+    "@angular-eslint/schematics": "^17.3.0",
+    "@angular-eslint/template-parser": "^17.3.0",
+    "@angular/cli": "^17.3.2",
+    "@angular/compiler-cli": "^17.3.2",
     "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
-    "@commitlint/cli": "^19.0.3",
-    "@commitlint/config-conventional": "^19.0.3",
+    "@commitlint/cli": "^19.2.1",
+    "@commitlint/config-conventional": "^19.1.0",
     "@types/jasmine": "^5.1.4",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^20.11.24",
-    "@typescript-eslint/eslint-plugin": "^7.1.0",
-    "@typescript-eslint/parser": "^7.1.0",
-    "@vendure/ngx-translate-extract": "^9.1.0",
-    "autoprefixer": "^10.4.18",
+    "@types/node": "^20.11.30",
+    "@typescript-eslint/eslint-plugin": "^7.4.0",
+    "@typescript-eslint/parser": "^7.4.0",
+    "@vendure/ngx-translate-extract": "^9.1.1",
+    "autoprefixer": "^10.4.19",
     "eslint": "^8.57.0",
     "husky": "^9.0.11",
     "jasmine-core": "^5.1.2",
@@ -75,15 +75,15 @@
     "karma-jasmine-html-reporter": "^2.1.0",
     "lint-staged": "^15.2.2",
     "ngx-devkit-builders": "^1.2.1",
-    "ngx-i18n-extract-regex-cli": "^0.0.7",
-    "postcss": "^8.4.35",
+    "ngx-i18n-extract-regex-cli": "^0.0.8",
+    "postcss": "^8.4.38",
     "prettier": "^3.2.5",
-    "stylelint": "^16.2.1",
+    "stylelint": "^16.3.1",
     "stylelint-config-standard": "^36.0.0",
     "stylelint-config-standard-scss": "^13.0.0",
     "stylelint-config-tailwindcss": "^0.0.7",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5.3.3"
+    "tailwindcss": "^3.4.3",
+    "typescript": "^5.4.3"
   },
   "engines": {
     "node": ">=20 <22",
