@@ -1,4 +1,4 @@
-import { Directive, EventEmitter, HostListener, Output, input } from '@angular/core';
+import { Directive, HostListener, input, output } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { TranslateService } from '@ngx-translate/core';
 import { filter, first, switchMap } from 'rxjs';
@@ -11,8 +11,7 @@ import { CustomConfirmDialog, CustomConfirmDialogService } from '../services/cus
 })
 export class PostDeleteDirective {
   public id = input.required<number>({ alias: 'appPostDelete' });
-
-  @Output() public deleted = new EventEmitter<number>();
+  public deleted = output<number>();
 
   constructor(
     private apiService: ApiService,
